Pull lint command implementation out, reword changelog.
